Move checkks into one file such that it can be used elsewhere
